refactor(icons): export IconRenderer
