+++ packages/icons/src/icon-renderer.js
